requirements.txt
django
ka-drf-utils==0.5.0
django-cors-headers
djangorestframework-simplejwt>=4.4.0,<5.0
django-filter>=2.4.0
django-upload-validator==1.1.5
django-extra-fields==3.0.2
Pillow==8.0.1
pyOpenSSL==19.1.0

# caches and queues
psycopg2-binary
kombu>=5.0.2,<6.0
msgpack==1.0.0


